Add operation `sleep' to sleep specific seconds and exit
It is designed to be used in custom commands.

Another thought is prefixing `@' to a custom command to make it a batch
command, so users can write `@ping localhost -n 6 >nul' in
`somecustomcmd.txt' to make it sleep 5 seconds between other 2 commands.
But executing arbitrary batch commands read from a file is dangerous.